Rozdelenie na menšie komponenty pre lepšiu spravovateľnosť - fileupload, select, návod pre dizajnérov, api route.js
